% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geommc-package.R
\docType{package}
\name{geommc-package}
\alias{geommc}
\alias{geommc-package}
\title{geommc: Geometric Markov Chain Sampling}
\description{
Simulates from discrete and continuous target distributions using geometric Metropolis-Hastings (MH) algorithms. Users specify the target distribution by an R function that evaluates the log un-normalized pdf or pmf. The package also contains a function implementing a specific geometric MH algorithm for performing high dimensional Bayesian variable selection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vroys/geommc}
}

}
\author{
\strong{Maintainer}: Vivekananda Roy \email{vroy@iastate.edu} (\href{https://orcid.org/0000-0002-2964-9503}{ORCID})

}
\keyword{internal}
